requirements/local.txt
-r base.txt

django-debug-toolbar==3.8.1
django-extensions==3.2.1

graphviz==0.20.1
pyparsing==3.1.1
pydot==1.4.2
pytest==7.4.2
